REF: refactor rm_rm into utils module
